pyproject.toml
[project]
name = "pdf-researcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.1.0",
    "langchain==0.3.27",
    "langchain-chroma==0.2.6",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.76",
    "langchain-google-genai==2.1.10",
    "langgraph==0.6.7",
    "logtail-python>=0.3.4",
    "nest-asyncio>=1.6.0",
    "pymupdf>=1.26.4",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
    "valkey>=6.1.1",
]
